Foldable Card: New prop clickableHeaderText that allows just the header text to be clickable
Sometimes, like in Jetpack when you already have a clickable item (toggle) in the header, you still want part of the header to be clickable and expand, but not the whole thing, as it would clash with the toggle.  This allows just the header text to be clickable.

Also adds the existing `clickableHeader` to the listed propTypes, as it was missing from there before.